Fixes #3509: UnRAR would not recognize files with international filenames
Root cause #1:
When opening a RAR file, filenames are converted to UTF-8. When opening a file
we need to convert the filenames to UTF-8 again before doing the string
comparison to decide if we've found the file.

Root cause #2:
Using CSmartStr in extract.cpp is only smart if we look at how they're used.
sizeof (foo) was returning the size of a pointer instead of the length of the
buffer.