Ticket #8661: xbmc-python-module-no-as-needed.patch
Add -Wl,--no-as-needed to the link line of python wrapper module. We use linking flag -Wl,--as-needed by default to avoid unneeded library dependencies. However, the python xbmc module does not use libpython even if it links against it, causing --as-needed to drop it, and thus xbmc can't resolve python symbols through the module. Therefore this patch explicitely disables it.