[76d7202] smartplaylists: add a library.smartplaylist property to a xsp-based item list (fixes xbmc#13412)
CFileItem::IsSmartPlaylist() only works based on the path and extension of a
CFileItem object. When retrieving a list of file items through
CSmartPlaylistDirectory we replace the original path (to the *.xsp) of the
list with the library/database-specific path of that list to be able to
use advanced filtering et. al. on these lists as well. Therefore we loose
the possibility to determine whether the list originates from a smartplaylist
or not. This results in the selection of a wrong CGUIViewState which does not
provide the 'Sort by Playlist' option and therefore any custom sorting
specified in the smartplaylist is ignored/lost. Therefore we add a property
'library.smartplaylist' to the CFileItemList object (similar to
'library.filter' for filter-based custom library nodes) and also check for
that in CFileItem::IsSmartPlaylist().