[5480331] media library: don't replace the path of m_vecItems with the filter path
In the current implementation upon retrieving a list of items in
CGUIMediaWindow::Update() we overwrite the path passed in as a parameter with
the path assigned to the new CFileItemList instance (the filter path). This
results in loss of important information because the path is used in many
places (e.g. getting the right CGUIViewState implementation). But for
advanced filtering to work we need to also be able to access the filter path
containing the active filter as a JSON representation in the filter URL
parameter. This allows to pass an active filter down to a sub-directory and
also to return back to a filtered list from a sub-directory.
To avoid overwriting the original path used to retrieve the list of items we
store the filter path in a new member variable m_strFilterPath instead and
also store the filter path of a list in the current path history to be able
to return to the filtered list from a sub-directory.