[af1ae88] Fixed calls to Event::Wait* that were supposed to be interruptible.
Durring the threading-refactor, there were several places where calls to CThreads::WaitForSingleObject was replaced with calls to the replacement for ::WaitForSingleObject. The CThreads version of the call would return when the thread was stopped. This fix puts back that behavior in the places that were it was originally.

Also incorporated elupus' comments on style and which approach to use.