[5b4f951] fixed: python crashes with multithreaded scripts in certain cases
We deinitialize the Python interpreter when no scripts are running.
However, the global interpreter lock (which is created by
PyEval_InitThreads()) is not unallocated. When we eventually
reinitialize the interpreter and call PyEval_InitThreads() to allocate
and lock the global interpreter lock, it doesn't actually do anything as
the lock is already allocated, and the lock is left in unlocked state,
which causes python threads to collide.

Fix that by checking if the lock already exists and simply acquire it in
that case.

Issue and fix confirmed on several Mandriva Linux x86_64 systems with
external Python 2.6.x.