[a2763c7] [WIN] fix crash on start due to not correctly setting thread name
Follow MS example closely, __try/__except is the way to catch the
exceptions when the thread doesn't have a handler to transcode to
regular C++ exceptions.
We don't have such a handler since the thread refactor and this is a
platform specific file, so no there is no problem using an MS language
extension.