[ccfb770] decouple rendering and window closing, add AfterRender method (that will be called each frame even if we actually didn't render anything) and move window closing logic there - including 446b52e
XBOX: After this change, OSD and Dialogs were not opening (rendering) during videoplayback in fullscreen. Solution was to call AfterRender() in CGUIWindowManager::RenderDialogs() which is called from Videoplayback
rendering thread. Hovever this still didn't fix main issues which we introduced with Dirty Regions (blank screen on GUIWindowFullscreen close, UI glitches when opening dialogs and general choppy navigation in dialogs).
We will need to investigate more XboxRenderer::Render and CApplication::Render. Kodi is rendering everything from CApplication::Render() but for Xbox that's not case. When we are in Fullscreen and videoplayback is not paused
we are using XboxRenderer::Render() to render everything what we see on screen